Fix causal_conv1d dtype mismatch: cast bias to match weight dtype
The restore_low_dim_params_to_fp32 function keeps conv1d bias in fp32
while weights are bf16. causal_conv1d_fn requires matching dtypes.

+++ train_gpt.py
@@ -770,7 +770,7 @@ class MambaBlock(nn.Module):
                 # Fused causal conv1d + SiLU
                 conv_weight = self.conv1d.weight.squeeze(1)  # (d_inner, d_conv)
                 x_in = _causal_conv1d_fn(x=x_in, weight=conv_weight,
-                                         bias=self.conv1d.bias, activation="silu")
+                                         bias=self.conv1d.bias.to(conv_weight.dtype), activation="silu")
             else:
                 x_in = self.conv1d(x_in)[:, :, :residual.size(1)]
                 x_in = F.silu(x_in)
